Extract DecoratingWorld from TracingWorld
